test(console): move the GError suite out of scope
`scripts/audit-runtimes.mjs` skips `*.gjs.spec.ts` — "specs are allowed to
exercise GJS-only paths" — and scans every other `.ts` in a package. With the
GError cases inside `index.spec.ts`, its `await import('gi://…')` counted as a
source signal and the audit reported @gjsify/console's declared triplet as
drifting to node:partial, browser:partial. That audit is a required check, so a
test made it red.

Same suite, same assertions: Node 168 and GJS 252, unchanged.

Also records two measured placement defects in status/open-todos.md that this
PR does NOT fix, so they do not ship silently: an adopted composite offsets by
its own internal children (a prepend into an AdwPreferencesPage renders in the
wrong order, at exit 0), and removing an element child does not restate the text
it displaced (a labelled Button goes blank).
