+++ packages/node/console/src/test.mts
@@ -1,5 +1,6 @@
 import { run } from '@gjsify/unit';
 
+import gerrorSuite from './gerror.gjs.spec.js';
 import testSuite from './index.spec.js';
 
-run({ testSuite });
+run({ testSuite, gerrorSuite });
